Move power tag definitions into an _impl/scaled subdir
- more in line with other tags (although this one doesn't have tag in the
  name
- tag and convert_operator definitions separated into different headers
